Extending the interface in order to add possibility to implement jumps. Now, the function objects mvr and mvd receive an additional integer argument e indicating the edge in the tree for which the function has been called. Models with jumps can use this index to check whether there was a jump at the beginning of the edge, e.g. xi[e] == 1
The change in the interface has been implemented in the OU model but has to be implemented in the other models as well.

Models implementing jumps should most likely have to add an argument e to the function returned by V(lambda, P, P_1, Sgima, threshold). This, however, is model specific and does not make part of the PCMBase interface.

Some formatting